services/orchestrator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-dotenv>=1.0.0
jsonpatch>=1.33
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pydantic>=2.7.0
ruff>=0.4.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
livekit-api>=0.8.0
# Optional: Redis session cache
redis>=5.0.0
# Optional: ChromaDB for RAG long-term memory
chromadb>=0.4.0

